chore(provd): rebase and move mock to locale_mock
